<!--
Thank you for working on PD! Please read PD's [CONTRIBUTING](https://github.com/pingcap/pd/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed
-->

### What problem does this PR solve? <!--add the issue link with summary if it exists-->
https://github.com/pingcap/pd/issues/1847

### What is changed and how it works?
Avoided slice re-allocation and copying cost
changed from histogram to summary for distribution information

### Release note <!-- bugfixes or new feature need a release note -->


### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Manual test (add detailed scripts or steps below)

Code changes
- No functionality change, improvement in performance
- For benchmark, added Prometheus code to print metrics data at the end of a run
 
Side effects

- In case of many clients created within single Tidb instance can lead to slightly higher memory consumption

Related changes

- PR to update `pingcap/docs`/`pingcap/docs-cn`:
- PR to update `pingcap/tidb-ansible`:
- Need to cherry-pick to the release branch