@nolouch 2020-03-30
Thanks for your contribution. :)
I have a real environment and reproduced it, could you test in it with your branch?
Let me contact you on slack.