@mgkanani 2020-04-04
Performance Comparison(may vary system to system):
For both request count/sec = 0.9M +- 10%
@nolouch , thank you for the environment.

Master Branch:
`
Total:
count:105642065, max:76, min:0, >1ms:0, >2ms:0, >5ms:9420055 (8.917%), >10ms:96057544 (90.927%), >30ms:164413 (0.156%)
`

With this PR's changes:
`
Total:
count:104146574, max:69, min:0, >1ms:2160 (0.002%), >2ms:32434 (0.031%), >5ms:8137420 (7.813%), >10ms:95903141 (92.085%), >30ms:71371 (0.069%)
`