@mgkanani 2020-04-12
Hi @nolouch ,@tiancaiamao,

There is substantial difference between benchmarks `pd-tso-bench` and `sysbench(with tidb-pd-tikv cluster)`:

- With pd-tso-bench: number of requests to number of clients ratio is higher. So batching size plays a role in it.  E.g. we used 10k clients and number of requests per second was ~1M. So per client, on avg ~100(=1M/10k) batching possible. 
- Same changes with sysbench: 2000 clients generates ~2k request so lesser chance for batching. With this experiment, batch size hardly goes beyond 256 and avg is 2-4. I tried profiling and client.tsLoop function executed for ~0.05%.

If sysbench is near real time then there is no point of setting `maxMergeTSORequests` way higher ~10k.
If possible, `pd_client_request_handle_tso_batch_size` metrics data from real-server with high load may help in setting the value. Based on the data, I will make relevant changes to [maxMergeTSORequests size](https://github.com/pingcap/pd/pull/2300/files#diff-1cd41a2611da49ce63630d5a51bd82c5R105) and [default Slice size](https://github.com/pingcap/pd/pull/2300/files#diff-1cd41a2611da49ce63630d5a51bd82c5R192)