@mgkanani 2020-04-12
> Hi @mgkanani , the environment running a TiDB cluster with a heavy workload, does your PR benefit for that situation?

Yes, if there are many pending requests which lead to higher the batch-size.